@@ -3,13 +3,14 @@ description: A new Flutter project.
 publish_to: 'none' 
 version: 1.0.0+1
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  cloud_firestore: ^2.1.0
+  cloud_firestore: ^0.16.0
   cupertino_icons: ^1.0.2
+  geoflutterfire: ^2.2.2
   google_maps_flutter: ^2.0.3
 
 dev_dependencies:
